Road and Citizen fixes
added colliders to the roads that weren't working.  Linked the SpawnNPCs script with the play button to spawn the NPCs when the play button is clicked.  Changed the SpawnNPCs script so that it will keep track of the number spawned and will only spawn new NPCs if the total population changes